LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-flexbox/intrinsic-size/row-use-cases-001-expected.txt

https://bugs.chromium.org/p/chromium/issues/detail?id=423112


first checkbox example

second checkbox example
https://bugs.chromium.org/p/chromium/issues/detail?id=464210

MMMMM
MMMMM
https://bugs.chromium.org/p/chromium/issues/detail?id=533627#c3

hello
hello
https://bugs.chromium.org/p/chromium/issues/detail?id=441768 author wants below blue block to have 0 width

https://bugs.chromium.org/p/chromium/issues/detail?id=1144312

phantom whitespace
also phantom whitespace
no more phantom whitespace
https://bugs.chromium.org/p/chromium/issues/detail?id=1154243

Test content
Test content

FAIL left < right assert_less_than_equal: expected a number less than or equal to 211.28125 but got 239.28125
FAIL no overflow assert_less_than_equal: expected a number less than or equal to 178.28125 but got 233
FAIL same widths assert_equals: expected 278 but got 52
FAIL zero width assert_equals: expected 0 but got 200
FAIL same heights assert_equals: expected 28 but got 40
FAIL no overflow 2 assert_less_than_equal: expected a number less than or equal to 183 but got 301
